+++ src/components/EquipmentDropdown/component.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { PropTypes } from 'prop-types';
 import { gearShape } from '../../helpers/proptypeShapes';
-import ButtonStandard from '../../helpers/buttons/ButtonStandard';
+import ButtonStandard from '../widgets/buttons/ButtonStandard';
 import { createArrayOfEquipment, filterEquipment, createFilterSet } from '../../helpers/equipmentListFunctions';
 import { addEquipment } from '../../helpers/actionHelpers';
 import './EquipmentDropdown.css';
