+++ src/components/WeaponsModalSelection/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import ButtonStandard from '../../helpers/buttons/ButtonStandard';
-import ButtonInfo from '../../helpers/buttons/ButtonInfo';
+import ButtonStandard from '../widgets/buttons/ButtonStandard';
+import ButtonInfo from '../widgets/buttons/ButtonInfo';
 
 const WeaponsModalSelection = ({ firearmsArray, closeShowFirearms, handleShowGunStats, handleAddFirearm }) => {
   const newFirearmsArray = firearmsArray;
